+++ src/components/Contentful/HomePage/NewHomeLoansByCategoryGrid.vue
@@ -39,11 +39,11 @@ const getContentfulLoanChannels = content => {
 
 const loanCategoryPrefetchQuery = gql`
 	${loanFieldsFragment}
-	query loanCategoryPrefetch($loanChannelIds: [Int]!) {
+	query loanCategoryPrefetch($loanChannelIds: [Int]!, $limit: Int) {
 		lend {
 			loanChannelsById(ids: $loanChannelIds) {
 				id
-				loans(limit: 1) {
+				loans(limit: $limit) {
 					values {
 						id
 						...loanFields
@@ -55,7 +55,7 @@ const loanCategoryPrefetchQuery = gql`
 
 export default {
 	name: 'NewHomeLoansByCategoryGrid',
-	inject: ['apollo', 'cookieStore'],
+	inject: ['apollo', 'cookieStore', 'device'],
 	components: {
 		KivaMultiCategoryGrid,
 		KvPageContainer,
@@ -102,13 +102,14 @@ export default {
 		}
 	},
 	apollo: {
-		preFetch(config, client, { content }) {
+		preFetch(config, client, { content, device }) {
 			const contentfulLoanChannels = getContentfulLoanChannels(content);
 			const id = contentfulLoanChannels[0]?.id;
 			return client.query({
 				query: loanCategoryPrefetchQuery,
 				variables: {
 					loanChannelIds: id ? [id] : [],
+					limit: device?.platform?.type === 'desktop' ? 6 : 1
 				},
 			});
 		},
@@ -116,12 +117,14 @@ export default {
 	created() {
 		// Fetch loan channel data from the cache
 		let data = {};
+		const isDesktop = this.device?.platform?.type === 'desktop';
 		try {
 			const id = this.contentfulLoanChannels[0]?.id;
 			data = this.apollo.readQuery({
 				query: loanCategoryPrefetchQuery,
 				variables: {
 					loanChannelIds: id ? [id] : [],
+					limit: isDesktop ? 6 : 1
 				},
 			});
 		} catch (e) {
@@ -130,11 +133,18 @@ export default {
 
 		// Create an array with placeholder loans for loading
 		const { loanLimit = 0 } = this.loanDisplaySettings;
-		const loanValues = Array(loanLimit).fill({ id: 0 });
 
 		// Get the fetched loan and merge it into the placeholder loan array
 		const loanChannel = data?.lend?.loanChannelsById[0] ?? { loans: { values: [] } };
-		loanValues[0] = loanChannel?.loans?.values[0];
+
+		let loanValues;
+		if (isDesktop) {
+			loanValues = loanChannel?.loans?.values;
+		} else {
+			loanValues = Array(loanLimit).fill({ id: 0 });
+			loanValues[0] = loanChannel?.loans?.values[0];
+		}
+
 		const loanChannelCopy = {
 			...loanChannel,
 			loans: {